// src/api/portaria/models/VehicleResponseDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export enum VehicleResponseDtoStatusEnum {
  INSIDE = 'INSIDE',
  OUTSIDE = 'OUTSIDE',
  FORCE_CLOSE = 'FORCE_CLOSE',
}

export type VehicleResponseDto = {
  /**
   * @type string
   */
  id: string;
  /**
   * @type string
   */
  placa: string;
  /**
   * @type string
   */
  marca: string;
  /**
   * @type string
   */
  modelo: string;
  /**
   * @type number
   */
  ano: number;
  /**
   * @type string | undefined
   */
  cor?: string;
  /**
   * @type string, date-time
   */
  createdAt: Date;
  /**
   * @type string, date-time
   */
  updatedAt: Date;
  /**
   * @type string
   */
  status: VehicleResponseDtoStatusEnum;
  /**
   * @type string | undefined
   */
  fotoUrl?: string;
};